deleted unused task functions and made all other task function tests work
