+++ src/TaskFunctions.tsx
@@ -72,59 +72,3 @@ export function delTask(task: Task, tasks: Task[]) {
     );
     return newTasks;
 }
-
-/*
-Modify Name attribute
-*/
-export function setName(newName: string, task: Task) {
-    return { ...task, name: newName };
-}
-
-/**
-Modify description attribute
- */
-export function setDescription(newDecsription: string, task: Task) {
-    return { ...task, description: newDecsription };
-}
-
-/**
-Modify status attribute
- */
-export function setStatus(newStatus: boolean, task: Task) {
-    return { ...task, status: newStatus };
-}
-
-/**
-Modify image attribute
- */
-export function setImage(newImage: string, task: Task) {
-    return { ...task, image: newImage };
-}
-
-/**
- function to modify the steps attribute
- */
-export function setSteps(newSteps: string[], task: Task) {
-    return { ...task, steps: newSteps };
-}
-
-/**
- function to modify the difficulty attribute
- */
-export function setDifficulty(newDifficulty: number, task: Task) {
-    return { ...task, difficulty: newDifficulty };
-}
-
-/**
- function to modify the number of users attribute 
- */
-export function setNumUsers(newNumUsers: number, task: Task) {
-    return { ...task, numUsers: newNumUsers };
-}
-
-/**
- * function to modify the time attribute
- */
-export function setTime(newTime: string, task: Task) {
-    return { ...task, time: newTime };
-}
